fix:修复czmoney的问题

+++ src/form/RecycleForm.h
@@ -18,7 +18,7 @@ void showRecycleConfirmForm(
     int                dimId,
     BlockSource&       region,
     int                slotIndex,
-    int                unitPrice,
+    double             unitPrice, // 修改为 double
     const std::string& commissionNbtStr = ""
 );
 void showRecycleShopManageForm(Player& player, BlockPos pos, int dimId, BlockSource& region);
