+++ src/form/ShopForm.h
@@ -12,6 +12,6 @@ void showShopChestItemsForm(Player& player, BlockPos pos, int dimId, BlockSource
 void showShopItemPriceForm(Player& player, const ItemStack& item, BlockPos pos, int dimId, BlockSource& region);
 void showShopItemManageForm(Player& player, const std::string& itemNbtStr, BlockPos pos, int dimId, BlockSource& region);
 void showShopChestManageForm(Player& player, BlockPos pos, int dimId, BlockSource& region);
-void showShopItemBuyForm(Player& player, const ItemStack& item, BlockPos pos, int dimId, int slot, int unitPrice, BlockSource& region, const std::string& itemNbtStr);
+void showShopItemBuyForm(Player& player, const ItemStack& item, BlockPos pos, int dimId, int slot, double unitPrice, BlockSource& region, const std::string& itemNbtStr);
 
 } // namespace CT
